+++ app/src/renderer/components/staking/PageBond.vue
@@ -1,10 +1,6 @@
 <template lang="pug">
 page.page-bond(title="Bond Atoms")
   div(slot="menu"): tool-bar
-    router-link(to='/staking')
-      i.material-icons arrow_back
-      .label Back
-
   part(:title="'Start bonding your '+ totalAtoms + ' ATOM'"): form-struct(
     :submit="onSubmit")
     .bond-group(:class="bondGroupClass(unbondedAtomsDelta)")
@@ -61,7 +57,10 @@ page.page-bond(title="Bond Atoms")
             type="number"
             placeholder="Atoms"
             step="1"
-            v-model.number="d.atoms")
+            min="0"
+            :max="$v.fields.delegates.$each[index].atoms.$params.between.max"
+            v-model.number="d.atoms"
+            @change.native="limitMax(d, $event)")
 
       form-msg(name="Atoms" type="required"
         v-if="!$v.fields.delegates.$each[index].atoms.required")
@@ -153,10 +152,6 @@ export default {
     oldUnbondedAtoms () {
       return this.totalAtoms - this.oldBondedAtoms
     },
-    // not used
-    // newBondedAtoms () {
-    //   return this.fields.delegates.reduce((sum, d) => sum + (d.atoms || 0), 0)
-    // },
     newUnbondedAtoms () {
       return this.fields.delegates.reduce((atoms, d) => {
         let delta = d.oldAtoms - d.atoms
@@ -210,8 +205,8 @@ export default {
         this.$store.commit('activateDelegation')
         try {
           await this.$store.dispatch('submitDelegation', this.fields)
-          this.$store.commit('notify', { title: 'Atoms Bonded',
-            body: 'You have successfully updated your delegations.' })
+          this.$store.commit('notify', { title: 'Successful Delegation',
+            body: 'You have successfully bonded / unbonded.' })
           this.$router.push('/staking')
         } catch (err) {
           this.$store.commit('notifyError', { title: 'Error While Bonding Atoms',
@@ -288,14 +283,17 @@ export default {
           restrictSize: { min: { width: offset } }
         })
         .on('resizemove', (event) => {
-          let target = event.target
-          let ratio = (event.rect.width - offset) / (this.bondBarOuterWidth - offset)
-          let rawAtoms = ratio * this.totalAtoms
-
-          target.style.width = event.rect.width + 'px'
-
-          this.updateDelegateAtoms(target.id.split('-')[1], rawAtoms)
+          this.handleResize(event.target, event.rect.width)
         })
+    },
+    handleResize (element, width) {
+      let offset = this.bondBarScrubWidth
+      let ratio = Math.round((width - offset) / (this.bondBarOuterWidth - offset) * 100) / 100
+      let rawAtoms = ratio * this.totalAtoms
+
+      element.style.width = width + 'px'
+
+      return this.updateDelegateAtoms(element.id.split('-')[1], rawAtoms)
     },
     updateDelegateAtoms (delegateId, rawAtoms) {
       let d = this.fields.delegates.find(d => d.id === delegateId)
@@ -305,8 +303,8 @@ export default {
         d.deltaAtoms = this.delta(rawAtoms, d.oldAtoms, 'int')
         d.deltaAtomsPercent =
           this.percent(this.delta(rawAtoms, d.oldAtoms), this.totalAtoms)
+        return d
       }
-      return d.atoms
     },
     setBondBarOuterWidth () {
       let outerBar = this.$el.querySelector('.bond-bar__outer')
@@ -329,6 +327,14 @@ export default {
         value = Math.round(ratio * 100)
       }
       return value + '%'
+    },
+    limitMax (delegate, event) {
+      let max = parseInt(event.target.max)
+      if (delegate.atoms >= max) {
+        console.log(`${delegate.atoms} <= ${max}`)
+        delegate.atoms = max
+        return
+      }
     }
   },
   async mounted () {