+++ app/src/renderer/components/staking/PageDelegates.vue
@@ -1,5 +1,5 @@
 <template lang="pug">
-page#page-delegates(title='Delegates')
+page(title='Validators and Candidates')
   div(slot="menu"): tool-bar
     a(@click='setSearch(true)')
       i.material-icons search
@@ -78,9 +78,10 @@ export default {
       order: 'desc',
       properties: [
         { title: 'Name', value: 'description.moniker', class: 'name' },
-        { title: 'Vote %', value: 'shares', class: 'percent_of_vote' },
-        { title: 'Votes', value: 'voting_power', class: 'number_of_votes' },
+        { title: '% of Vote', value: 'shares', class: 'percent_of_vote' },
+        { title: 'Total Votes', value: 'voting_power', class: 'number_of_votes' },
         { title: 'Your Votes', value: 'bonded', class: 'bonded_by_you' },
+        { title: 'Status', value: 'status', class: 'status' },
         { title: '', value: '', class: 'action' }
       ]
     }
@@ -97,10 +98,10 @@ export default {
     },
     setSearch (bool) { this.$store.commit('setSearchVisible', ['delegates', bool]) }
   },
-  mounted () {
+  async mounted () {
     Mousetrap.bind(['command+f', 'ctrl+f'], () => this.setSearch(true))
     Mousetrap.bind('esc', () => this.setSearch(false))
-    this.updateDelegates(this.user.address)
+    await this.updateDelegates(this.user.address)
   }
 }
 </script>
@@ -108,7 +109,7 @@ export default {
 @require '~variables'
 
 .delegates-container
-  padding-bottom 6rem
+  padding-bottom 3rem
 
 .fixed-button-bar
   padding 0.5rem 1rem
@@ -120,6 +121,7 @@ export default {
   left 0
   right 0
   z-index z(toolBar)
+  
   .label
     color bright
     line-height 2rem
@@ -127,8 +129,7 @@ export default {
       font-weight bold
 
 @media screen and (min-width: 768px)
-  .delegates-container
-    padding-bottom 7rem
+    padding-bottom 4rem
 
   .fixed-button-bar
     padding 1rem