+++ test/e2e/wallet.js
@@ -1,42 +1,13 @@
 let { spawn } = require('child_process')
 let test = require('tape-promise/tape')
 let launchApp = require('./launch.js')
-let { navigate, newTempDir, waitForText, sleep } = require('./common.js')
+let { navigate, newTempDir, waitForText, sleep, login, logout } = require('./common.js')
 
-function cliInit (t, home) {
-  t.test('basecli init', function (t) {
-    let child = spawn('basecli', [
-      'init',
-      '--home', home,
-      '--chain-id', 'localtestnet',
-      '--node', 'localhost:46657'
-    ])
-    child.stdout.once('data', () => {
-      child.stdin.write('y\n')
-      child.once('exit', (code) => {
-        t.equal(code, 0, 'exited with exit code 0')
-        t.end()
-      })
-    })
-  })
-
-  t.test('recover basecli test key', function (t) {
-    let child = spawn('basecli', [
-      'keys', 'recover', 'testkey',
-      '--home', home
-    ])
-    child.stdin.write('1234567890\n')
-    child.stdin.write('chair govern physical divorce tape movie slam field gloom process pen universe allow pyramid private ability\n')
-    child.once('exit', (code) => {
-      t.equal(code, 0, 'exited with exit code 0')
-      t.end()
-    })
-  })
-}
+let binary = process.env.BINARY_PATH
 
 function cliSendCoins (home, to, amount) {
-  let child = spawn('basecli', [
-    'tx', 'send',
+  let child = spawn(binary, [
+    'client', 'tx', 'send',
     '--name', 'testkey',
     '--to', to,
     '--amount', amount,
@@ -50,107 +21,110 @@ function cliSendCoins (home, to, amount) {
 }
 
 test('wallet', async function (t) {
-  let app, client
+  let {app, home} = await launchApp(t)
+  let client = app.client
   let $ = (...args) => client.$(...args)
-  app = await launchApp(t)
-  client = app.client
-
-  let cliHome = newTempDir()
-  console.error(`basecli home: ${cliHome}`)
-  cliInit(t, cliHome)
+  
+  await logout(client)
+  await login(client, 'testkey')
 
   let balanceEl = (denom) =>
-    $(`div=${denom.toUpperCase()}`)
+    $(`//div[contains(text(), "${denom.toUpperCase()}")]`)
       .$('..')
       .$('div.ni-li-dd')
 
-  t.test('receive', async function (t) {
-    navigate(t, client, 'Balances')
+  t.test('send', async function (t) {
+    async function goToSendPage () {
+      await navigate(client, 'Balances')
+      await $('.ni-li-dt=FERMION').$('..').$('..').click()
+    }
 
-    let address
-    t.test('address', async function (t) {
-      let addressEl = $('div=Address').$('..').$('div.ni-li-dd')
-      address = await addressEl.getText()
-      t.ok(address, `address: ${address}`)
-      t.equal(address.length, 40, 'address is correct length')
+    await navigate(client, 'Balances')
+
+    let sendBtn = () => $('.ni-form-footer button')
+    let addressInput = () => $('#send-address')
+    let amountInput = () => $('#send-amount')
+    let denomBtn = (denom) => $(`option=${denom.toUpperCase()}`)
+
+    t.test('fermion balance before sending', async function (t) {
+      let fermionEl = balanceEl('fermion')
+      let balance = await fermionEl.getText()
+      t.equal(balance, '9007199254740992', 'fermion balance is correct') 
       t.end()
     })
 
-    t.test('mycoin balance before receiving', async function (t) {
-      let mycoinEl = balanceEl('mycoin')
-      let balance = await mycoinEl.getText()
-      t.equal(balance, '0', 'mycoin balance is 0')
+    t.test('hit send with empty form', async function (t) {
+      await goToSendPage()
+      await sendBtn().click()
+      t.equal(await sendBtn().getText(), 'Send Tokens', 'not sending')
       t.end()
     })
 
-    t.test('receive mycoin', async function (t) {
-      await cliSendCoins(cliHome, address, '1000mycoin')
+    t.test('address w/ less than 40 chars', async function (t) {
+      await goToSendPage()
+      await addressInput().setValue('012345')
+      await $('div=Address must be exactly 40 characters').waitForExist()
+      t.pass('got correct error message')
+      await sendBtn().click()
+      t.equal(await sendBtn().getText(), 'Send Tokens', 'not sending')
       t.end()
     })
 
-    t.test('mycoin balance after receiving', async function (t) {
-      let mycoinEl = () => balanceEl('mycoin')
-      await waitForText(mycoinEl, '1000', 8000)
-      t.pass('received mycoin transaction')
+    t.test('address w/ 40 chars', async function (t) {
+      await goToSendPage()
+      await addressInput().setValue('0'.repeat(40))
+      t.notOk(await client.isExisting('div=Address must be exactly 40 characters'), 'no error message')
+      await sendBtn().click()
+      t.equal(await sendBtn().getText(), 'Send Tokens', 'not sending')
+      t.end()
+    })
+
+    t.test('amount set', async function (t) {
+      await goToSendPage()
+      await amountInput().setValue('100')
+      await sendBtn().click()
+      t.equal(await sendBtn().getText(), 'Send Tokens', 'not sending')
+
+      t.end()
+    })
+
+    t.test('send', async function (t) {
+      await goToSendPage()
+      await amountInput().setValue('100')
+      await addressInput().setValue('3F52AFC4FB737A0296EFE331885FCC476980B3BD')
+      await sendBtn().click()
+      await client.waitForExist('.ni-notification', 5000)
+      let msg = await client.$('.ni-notification .body').getText()
+      t.ok(msg.includes('Success'), 'Send successful')
+      
+      t.end()
+    })
+
+    t.test('own balance updated', async function (t) {
+      await navigate(client, 'Balances')
+
+      // TODO should not be necessary
+      await sleep(1000)
+      await client.$('.material-icons=refresh').click()
+
+      let mycoinEl = () => balanceEl('fermion')
+      await waitForText(mycoinEl, '9007199254740892')
+      t.pass('balance is reduced by 100')
       t.end()
     })
 
     t.end()
   })
 
-  t.test('send', async function (t) {
-    navigate(t, client, 'Send', 'Send Coins')
+  t.test('receive', async function (t) {
+    t.test('fermion balance after receiving', async function (t) {
+      await logout(client)
+      await login(client, 'testreceiver')
+      await navigate(client, 'Balances')
 
-    let sendBtn = () => $('button=Send Now')
-    let addressInput = () => $('#send-address')
-    let amountInput = () => $('#send-amount')
-    let denomBtn = (denom) => $(`button=${denom.toUpperCase()}`)
-
-    t.test('hit send with empty form', async function (t) {
-      await sendBtn().click()
-      t.equal(await sendBtn().getText(), 'Send Now', 'not sending')
-      t.end()
-    })
-
-    t.test('address w/ less than 40 chars', async function (t) {
-      await addressInput().setValue('012345')
-      await $('div=Address must be exactly 40 characters').waitForExist()
-      t.pass('got correct error message')
-      await sendBtn().click()
-      t.equal(await sendBtn().getText(), 'Send Now', 'not sending')
-      t.end()
-    })
-
-    t.test('address w/ 40 chars', async function (t) {
-      await addressInput().setValue('0'.repeat(40))
-      t.notOk(await client.isExisting('div=Address must be exactly 40 characters'), 'no error message')
-      await sendBtn().click()
-      t.equal(await sendBtn().getText(), 'Send Now', 'not sending')
-      t.end()
-    })
-
-    t.test('amount set', async function (t) {
-      await amountInput().setValue('100')
-      await sendBtn().click()
-      t.equal(await sendBtn().getText(), 'Send Now', 'not sending')
-      // await $('div=Denomination is required').waitForExist()
-      // t.pass('got correct error message')
-      t.end()
-    })
-
-    t.test('denom set', async function (t) {
-      await denomBtn('mycoin').click()
-      await sleep(100)
-      await sendBtn().click()
-      t.end()
-    })
-
-    navigate(t, client, 'Balances')
-
-    t.test('own balance updated', async function (t) {
-      let mycoinEl = () => balanceEl('mycoin')
-      await waitForText(mycoinEl, '900')
-      t.pass('balance is now 900')
+      let fermionEl = () => balanceEl('fermion')
+      await waitForText(fermionEl, '100', 5000)
+      t.pass('received mycoin transaction')
       t.end()
     })