+++ test/unit/specs/node.spec.js
@@ -14,8 +14,6 @@ describe('LCD Connector', () => {
         destroy () {}
       }
     }))
-
-    process.env.COSMOS_UI_ONLY = 'false'
   })
 
   beforeAll(() => {
@@ -30,12 +28,6 @@ describe('LCD Connector', () => {
   it('should provide the nodeIP', () => {
     let node = LCDConnector('1.1.1.1')
     expect(node.nodeIP).toBe('1.1.1.1')
-  })
-
-  it('should return a mockClient if setting COSMOS_UI_ONLY', () => {
-    process.env.COSMOS_UI_ONLY = 'true'
-    let node = LCDConnector('1.1.1.1')
-    expect(node.generateKey().key.address).toBe('UI_ONLY_MODE')
   })
 
   it('should init the rpc connection on initialization', () => {